requirements.txt
alembic==1.4.2
aniso8601==8.0.0
APScheduler==3.6.3
attrs==19.3.0
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
coverage==5.1
elasticsearch==7.6.0
Flask==1.1.2
Flask-Ext==0.1
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.2
MarkupSafe==1.1.1
mimerender==0.6.0
pkg-resources==0.0.0
PyJWT==1.7.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-mimeparse==1.6.0
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
tzlocal==2.0.0
urllib3==1.25.9
Werkzeug==0.16.0
zipp==1.2.0
